pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rffl-boxscores"
version = "0.1.0"
description = "Clean ESPN fantasy boxscore exporter + validator"
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Kyle Thorsen" }]
dependencies = [
    "espn_api>=0.39.1",
    "pandas>=2.2.0", 
    "python-dateutil>=2.9.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
rffl-bs = "rffl_boxscores.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
]

[tool.setuptools]
packages = ["rffl_boxscores"]
include-package-data = false
